% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_homoplasy_indices.r
\name{calculate_homoplasy_indices}
\alias{calculate_homoplasy_indices}
\title{Calculates common homoplasy metrics for a set of phylogenetic tree(s)}
\usage{
calculate_homoplasy_indices(
  trees,
  cladistic_matrix,
  inapplicables_as_missing = FALSE,
  polymorphism_behaviour,
  uncertainty_behaviour,
  polymorphism_geometry,
  polymorphism_distance,
  state_ages,
  dollo_penalty
)
}
\arguments{
\item{trees}{A tree (\code{phylo} object) or set of trees (\code{multiPhylo} object).}

\item{cladistic_matrix}{A character-taxon matrix in the format imported by \link{read_nexus_matrix}. These should be discrete with rownames (taxon labels) matching the tip labels of \code{trees}.}

\item{inapplicables_as_missing}{Logical that decides whether or not to treat inapplicables as missing (TRUE) or not (FALSE, the default and recommended option).}

\item{polymorphism_behaviour}{One of either "missing", "uncertainty", "polymorphism", or "random". See details.}

\item{uncertainty_behaviour}{One of either "missing", "uncertainty", "polymorphism", or "random". See details.}

\item{polymorphism_geometry}{Argument passed to \link{make_costmatrix}.}

\item{polymorphism_distance}{Argument passed to \link{make_costmatrix}.}

\item{state_ages}{Argument passed to \link{make_costmatrix}.}

\item{dollo_penalty}{Argument passed to \link{make_costmatrix}.}
}
\value{
Text.
}
\description{
Given a tree, or set of trees, and a cladistic matrix calculates common homoplasy metrics for each tree.
}
\details{
Several metrics have been developed to assess the degree of homoplasy - broadly the proportion of parallelisms plus reversals to first acquisitions of derived states - implied by a tree. This function calculates the common forms of these, excluding those that require computationally intensive permutations to which R is not well suited.

\bold{Basic terms}

Common metrics are built around the minimum number of steps - or really the minimum cost as steps refers to only the discrete case - a character can have on the tree (m), the maximum number (g), and those sampled from the empirical topolog(ies) (s). When summed across all characters these are given as the capitals M, G, and S, respectively (Kitching et al. 1998). The function also tabulates h/H which is given as s - m/S - M (for a character or entire matrix, respectively), and was proposed in Hoyal Cuthill (2015) and is identical to Fisher's (1994) "parsimony debt".

\emph{Minimum steps (cost)}

The minimum number of steps (cost) is given by either the length of the minimum spanning tree for a symmetric character, or the shortest arboresence for an asymmetric character, on the costmatrix of \emph{sampled} tip states. More specifically, although many taxa may be involved, we can imagine for the minimum possible number of steps (cost) any taxa that share a character state must form a subtree and that any arrangement of that subtree would result in zero inferred steps (cost). Thus each such subtree may be collapsed to a single node (vertex) representing that character state and hence the problem collapses to a minimum spanning tree on a costmatrix (i.e., a graph representation) problem. Fortunately, such graph theory problems have wide applicability and so a range of solutions are already available in the literature. For example, for a symmetric costmatrix - the case of an undirected graph - a minimum spanning tree can be found with Borůvka's (Borůvka 1926a,b; Choquet 1938; Florek et al. 1951; Nešetřil et al. 2001), DJP's (Jarnik 1930; Prim 1957; Dijkstra 1959), or Kruskal's (Kruskal 1956) algorithm. For an asymmetric costmatrix the problem is the same as finding a minimum spanning tree on a directed graph (digraph) and because such a tree must have a tree-like shape (including a root and subsequent direction of branching) this is more technically called an arboresence. Again, though, the wide applicablity of solutions means multiple algorithms for finding such a tree are already available (e.g., Chu and Liu 1965; Edmonds 1967; Bock 1971).

\emph{Maximum steps (cost)}

Mathematical proofs for maximum tree lengths are considerably more complex to solve (Hoyal Cuthill et al. 2010), and to the best of my knowledge there is no general costmatrix solution currently available. Instead, this function simply uses the approach common to other software and uses the minimum steps required on the star phylogeny as the maximum steps value. Note: this doesn't mean this is not the true maximum value, simply that this is not known for certain beyond the simpler cases, such as binary and unordered characters (see Hoyal Cuthill et al. 2010 and references therein).

\bold{The Consistency Index}

CI = M/S character version vs ensemble version (Kitching et al. 1998) [WITHOUT UNINFORMATIVE CHARACTERS? - RESCALED?]

\bold{The Retention Index}

Farris (1989) character version vs ensemble version (Kitching et al. 1998)

\bold{Permutation methods}

Cannot do HER as this requires inferring additional trees which Claddis is not setup to do. [REFS of alternates such as Hoyal Cuthill 2015]
}
\examples{

# TO DO

}
\references{
Bock, F., 1971. An algorithm to construct a minimum directed spanning tree in a directed Network. In, \emph{Developments in Operations Research}. Gordon and Breach, New York, pp. 29-44.

Borůvka, O., 1926a. O jistém problému minimálním [About a certain minimal problem]. \emph{Práce Moravské Přírodovědecké Společnosti V Brně III}, \bold{3}, 37-58.

Borůvka, O., 1926b. Příspěvek k řešení otázky ekonomické stavby elektrovodních sítí [Contribution to the solution of a problem of economical construction of electrical networks]. \emph{Elektronický Obzor}, \bold{15}, 153-154.

Choquet, G., 1938. Étude de certains réseaux de routes. \emph{Comptes Rendus de l'Académie des Sciences}, \bold{206}, 310-313.

Chu, Y. J. and Liu, T. H., 1965. On the shortest arborescence of a directed graph. \emph{Science Sinica}, \bold{14}, 1396-1400.

Dijkstra, E. W., 1959. A note on two problems in connexion with graphs. \emph{Numerische Mathematik}, \bold{1}, 269–271.

Edmonds, J., 1967. Optimum branchings. \emph{Journal of Research of the National Bureau of Standards Section B}, \bold{71B}, 233-240.

Farris, J. S., 1989. The retention index and the rescaled consistency index.\emph{Cladistics}, \bold{5}, 417-419.

Fisher, D. C., 1994. Stratocladistics: morphological and temporal patterns and their relation to phylogenetic process. In L. Grande and O. Rieppel (eds.), \emph{Interpreting the Hierarchy of Nature}. Academic Press, San Diego. pp133–171.

Florek, K., Łukaszewicz, J., Perkal, J., Steinhaus, H. and Zubrzycki, S., 1951. Sur la liaison et la division des points d'un ensemble fini. \emph{Colloquium Mathematicae}, \bold{2}, 282-285.

Hoyal Cuthill, J., 2015. The size of the character state space affects the occurrence and detection of homoplasy: modelling the probability of incompatibility for unordered phylogenetic characters. \emph{Journal ofTheoretical Biology}, \bold{366}, 24-32.

Hoyal Cuthill, J. F., Braddy, S. J. and Donoghue, P. C. J., 2010. A formula for maximum possible steps in multistate characters: isolating matrix parameter effects on measures of evolutionary convergence. \emph{Cladistics}, \bold{26}, 98-102.

Jarník, V., 1930, O jistém problému minimálním [About a certain minimal problem]. \emph{Práce Moravské Přírodovědecké Společnosti}, \bold{6}, 57–63.

Kitching, I. J., Forey, P. L., Humphries, C. J. and Williams, D. M., 1998. \emph{Cladistics: The Theory and Practice of Parsimony Analysis (Second Edition)}. Oxford University Press, Oxford. 229pp.

Kruskal, J. B., 1956. On the shortest spanning subtree of a graph and the traveling salesman problem. \emph{Proceedings of the American Mathematical Society}, \bold{7}, 48-50.

Nešetřil, J., Milková, E. and Nešetřilová, H., 2001. Otakar Borůvka on minimum spanning tree problem: translation of both the 1926 papers, comments, history. \emph{Discrete Mathematics}, \bold{233}, 3-36.

Prim, R. C., 1957. Shortest connection networks and some generalizations, \emph{Bell System Technical Journal}, \bold{36}, 1389-1401.
}
\seealso{
\link{calculate_tree_length}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
