% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tree_length.r
\name{calculate_tree_length}
\alias{calculate_tree_length}
\title{Calculates the parsimony length of a set of phylogenetic tree(s)}
\usage{
calculate_tree_length(
  trees,
  cladistic_matrix,
  inapplicables_as_missing = FALSE,
  include_polymorphisms,
  polymorphism_shape,
  polymorphism_distance,
  state_ages,
  dollo_penalty
)
}
\arguments{
\item{trees}{A tree (phylo object) or set of trees (multiPhylo object).}

\item{cladistic_matrix}{A character-taxon matrix in the format imported by \link{read_nexus_matrix}. These should be discrete with labels matching the tip labels of \code{tree}.}

\item{inapplicables_as_missing}{Logical that decides whether or not to treat inapplicables as missing (TRUE) or not (FALSE, the default and recommended option).}

\item{include_polymorphisms}{Argument passed to \link{make_stepmatrix}.}

\item{polymorphism_shape}{Argument passed to \link{make_stepmatrix}.}

\item{polymorphism_distance}{Argument passed to \link{make_stepmatrix}.}

\item{state_ages}{Argument passed to \link{make_stepmatrix}.}

\item{dollo_penalty}{Argument passed to \link{make_stepmatrix}.}
}
\value{
A list with multiple components, including:

\item{length}{The tree length (number of steps).}
\item{most_parsimonious_reconstructions}{A matrix where rows correspond to \emph{all} nodes (i.e., terminal and internal) in the order numbered by \code{ape}, and columns correspond to every unique most parsimonious reconstruction. I.e., if there is only one most parsimonious reconstruction there will be only one column.}
\item{input_tree}{The input topology used.}
}
\description{
Given a tree, or set of trees, and a cladistic matrix returns their parsimony length in number of steps.
}
\details{
Under the maximum parsimony criterion, a phylogenetic hypothesis is considered optimal if it requires the fewest number of evolutionary "steps". In order to evalulate this criterion we must therefore be able to calculate a tree's "length" (total steps assuming the fewest possible steps for every character used). Given a set of phylogenetic hypothes(es) and a cladistic matrix this function calculates the minimum length of each tree.

\bold{Input data format}

This function operates on a phylogenetic tree, or trees (in \code{ape} format), and a cladistic matrix (in \link{cladisticMatrix} format [UPDATE]). However, the algorithm used is based on the genralised stepmatrix approach of Swofford and Maddison (1992) and hence stepmatrices need to be defined for each character (this is done internally by calling \link{make_stepmatrix}), and some of the options are merely passed to this function.

\bold{Algorithm}

Technically the Swofford and Maddison (1992) algorithm is designed for ancestral state reconstruction, but as its' first pass of the tree assigns lengths for each possible state at each node the minimum value at the root is the tree length and hence by skipping the later steps this can be used as a tree length algorithm. The choice of the Swofford and Maddison algorithm, rather than the Wagner or Fitch algorithms (for ordered and unordered characters, respectively) is to generalize to the broadest range of character types, including asymmetric characters (Camin-Sokal, Dollo, stratigraphic), custom character types (specified using stepmatrices or character state trees), as well as to any resolution of tree (i.e., including multifurcating trees). The only restriction here is that the tree must be rooted such that time's arrow is explicitly present. This is essential as the root defines the lengths across the whole tree, but also for asymmetric characters directionality must be explicit. The two obvious downsides to this algorithm is that it can be slower and that it is not appropriate for unrooted trees.

\bold{Stepmatrices and stepmatrix options}

Stepmatrices are described in detail in the \link{make_stepmatrix} manual, as are the options that are passed from this function to that one.
Can be used to infer ancestral states or get tree lengths (could theoretically build a tree inference procedure around this by proposing different topologies, but ikely to be much slower than compiled software). Also a first step to generating character maps that can be used for plotting changes on a tree, performing rate analysis (see \link{test_rates}), or generating phylomorphospaces (not implemented yet).

\emph{Explicit options}

- inapplicables_as_missing = FALSE, what to do with inapplicable characters. Currently treat as missing is only real option, but can be differet state that is also inferred as an ancestor.
- More for uncertianties and polymorphisms (see below).
- Add option to predict missing tip values (but definitely not recommended! - see Lloyd 2018)
- Collpasing to a single output with ACCTRAN etc.?

\bold{Implementation}

Algorithms for reconstructing ancestral states under parsimony can be specific (e.g., for ordered characters only) or general (allowing for any character type, including highly specific custom-formats). The advantage of the former is usually that they are faster as multiple assumptions are already built in, but the latter offer more flexibility and represent the version applied here.

More specifically this function is based on the approach described by Swofford and Maddison (1992) that assumes a rooted tree and uses a stepmatrix to define the costs (in number of steps) of specific transitions, i.e., from state X to state Y. The rooting is key here as it sets the directionality of evolution (i.e., time) and hence allows the use of asymmetric characters (where the cost of going from state X to state Y is different than going from stete Y to state X) as well as other direction-based optimisations (e.g., ACCTRAN and DELTRAN).


\bold{Special character types}
\emph{Missing values}
\emph{Inapplicables}
\emph{Uncertainties}

In Claddis, uncertainties are coded by separating states with the slash character, e.g., 0/1. The Swofford and Maddison (1992) algorithm - and the implementation used here - treats uncertainties (automatically) by allowing any of the possible states at the tip (here 0 or 1, but not, say, 2 or 3). In other words, it will opt for solutions that effectively prefer whichever state is easiest (fewest steps) to reach from a proposed ancestral value. This could therefore be used to predict what the "true" value may be, at least under the assumption of parsimony. However, such phylogenetic "prediction" [OPTION?] is generally advised against (Lloyd 2018).

\emph{Polymorphisms}

TEXT?

\bold{Polytomies}

Another advantage of the Swofford and Maddison (1992) approach is that it does not require fully-bifurcating topologies, hence these are "allowed" as input. However, it is important to note that the implementation here will treat any such polytomies (also known as multifurcations) as "hard". I.e., that they genuinely represent one species splitting into three or more descendant species. As always, the Ian Malcolm caveat applies - "your scientists were so preoccupied with whether or not they sould, they didn’t stop to think if they should."


# Option to fix root state (polarize) somehow
# From is row, to is column (directional hence requires/assumes rooted trees)
# Record in manual that weighted parsimony is folly, but could use to chose between MPRs (maybe?)
\bold{Polymorphic characters}

# MANUAL: generalized (stepmatrix) solution as accounts for more possibilities
# MANUAL: really only for rooted trees (unrooted possible using same methods (see Swofford and Maddison (1992), but unclear why you would want to do this with trees without direction of time (i.e., a root).
# MANUAL: works with multifurcations and doesn't require a fully bifurcating tree, but does assume polytomies are therefore "hard".
# MANUAL: Swofford and Maddison (1992; p210): "Both [ACCTRAN and DELTRAN] start by either assuming that the ancestral state is known or by choosing a state from the set of optimal assignments to the root node" - i.e., always an arbitrainess problem at root!
MISSING DATA (NA) IS DEALT WITH BY SETTING ALL TIP VALUES FOR A MISSING STATE TO ZERO (AS PER SWOFFORD AND MADDISON 1992).
UNCERTAIN STATES (E.G., 0/1) CONSTRAIN ANCESTRAL STATES SLIGHTLY MORE THAN NA (ASSUMING THEY EXCLUDE SOME STATES. IMPLEMENTATION SAME AS SWOFFORD AND MADDISON (1992).
CALCULATES TREE LENGTHS!
}
\examples{

# Calculate maximum parsimony tree length for Gauthier 1986:
cladistic_matrix <- Claddis::gauthier_1986
tree <- ape::read.tree(
  text = "(Outgroup,(Ornithischia,(Sauropodomorpha,(Ceratosauria,Procompsognathus,
  Liliensternus,(Carnosauria,(Ornithmimidae,Saurornitholestes,Hulsanpes,(Coelurus,
  Elmisauridae,(Compsognathus,(Ornitholestes,Microvenator,Caenagnathidae,
  (Deinonychosauria,Avialae))))))))));"
)
calculate_tree_length(
  tree = tree,
  cladistic_matrix = cladistic_matrix,
  inapplicables_as_missing = FALSE,
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle",
  state_ages = c(),
  dollo_penalty = 999
)

}
\references{
Lloyd, G. T., 2018. Journeys through discrete-character morphospace: synthesizing phylogeny, tempo, and disparity. \emph{Palaeontology}, \bold{61}, 637-645.

Swofford, D. L. and Maddison, W. P., 1992. Parsimony, character-state reconstructions, and evolutionary inferences. \emph{In} R. L. Mayden (ed.) Systematics, Historical Ecology, and North American Freshwater Fishes. Stanford University Press, Stanford, p187-223.
}
\seealso{
\link{make_stepmatrix}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
