% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cladisticMatrix.R
\name{check_cladisticMatrix}
\alias{check_cladisticMatrix}
\title{Check cladisticMatrix object for errors}
\usage{
check_cladisticMatrix(cladistic_matrix)
}
\arguments{
\item{cladistic_matrix}{An object of class \code{cladisticMatrix}.}
}
\value{
An error message or empty vector if no errors found.
}
\description{
Internal function to check cladisticMatrix object for errors.
}
\details{
Internal Claddis function. Nothing to see here. Carry on.
}
\examples{

# Check that this is a valid cladisticMatrix object (will return error message as class
# is not set):
check_cladisticMatrix(cladistic_matrix = day_2016)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
