% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxonGroups.R
\name{check_taxonGroups}
\alias{check_taxonGroups}
\title{Check taxonGroups object for errors}
\usage{
check_taxonGroups(taxon_groups)
}
\arguments{
\item{taxon_groups}{An object of class \code{taxonGroups}.}
}
\value{
An error message or empty vector if no errors found.
}
\description{
Internal function to check taxonGroups object for errors.
}
\details{
Internal Claddis function. Nothing to see here. Carry on.
}
\examples{

# Create a taxon groups object:
taxon_groups <- list(
  Group_A = c("Species_1", "Species_2", "Species_3"),
  Group_B = c("Species_3", "Species_4"),
  Group_C = c("Species_5", "Species_6", "Species_7", "Species_8")
)

# Check that this is a valid taxonGroups object (will return error message as class
# is not set):
check_taxonGroups(taxon_groups = taxon_groups)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
