% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChronoPhyloMorphospacePlot.R
\name{ChronoPhyloMorphospacePlot}
\alias{ChronoPhyloMorphospacePlot}
\title{Chronophylomorphospace Plot}
\usage{
ChronoPhyloMorphospacePlot(pcoa_data, x_axis = 1, y_axis = 2,
  shadow = TRUE)
}
\arguments{
\item{pcoa_data}{Text.}

\item{x_axis}{Text.}

\item{y_axis}{Text.}

\item{shadow}{Text.}
}
\description{
Plots a 3D chronophylomorphospace.
}
\details{
Text.
}
\examples{
# Set random seed:
set.seed(4)

# Generate a random tree for the Michaux 1989 data set:
tree <- rtree(length(rownames(Michaux1989$matrix)))

# Set root time so latest tip terminates at the present:
tree$root.time <- max(diag(vcv(tree)))

# Add taxon names to the tree:
tree$tip.label <- rownames(Michaux1989$matrix)

# Perform a phylogenetic Principal Coordinates Analysis:
pcoa_data <- MorphMatrix2PCoA(Michaux1989, tree = tree)

# Plot a chronophylomorphospace:
ChronoPhyloMorphospacePlot(pcoa_data)
}
\author{
Emma Sherratt \email{emma.sherratt@gmail.com} and Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\references{
Sakamoto and Ruta
}
\keyword{coordinates}
\keyword{principal}

