% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_costmatrix.R
\name{classify_costmatrix}
\alias{classify_costmatrix}
\title{Classify a costmatrix character}
\usage{
classify_costmatrix(costmatrix)
}
\arguments{
\item{costmatrix}{An object of class \code{costMatrix}.}
}
\value{
A vector of named edges (X->Y) with their distances. The sum of this vector is the length of the minimum spanning tree.
}
\description{
Given a costmatrix, classifies it as one of twelve distinct character types.
}
\details{
Type I - constant (any type but must be one state)
Type II - symmetric binary (any type but must be two states)
Type III - multistate unordered (three plus states and unordered type)
Type IV - linear ordered symmetric (three plus states and ordered)
Type V - Non-linear ordered symmetric (four plus states and custom direct costs all equal)
Type VI - Binary irreversible (two states and irreversible type)
Type VII - Multistate irreversible (three states and irreversible type)
Type VIII - Binary Dollo (two states and Dollo type)
Type IX - Multistate Dollo (three plus states and Dollo type)
Type X - Multistate custom symmetric (three plus states, symmetric but variable direct transition costs)
Type XI - Binary custom asymmetric (two states, custom type costs unequal and non-infinite)
Type XII - Multistate custom asymmetric (three plus states, custom type, asymmetric costs and not Type VII)

# For whole cladistic data set:
Stratigraphic is something else(?)
Type XIII - Gap weighted (or is this just a linear ordered? I guess issue is when N states > N tips?)
Type XIV - Continuous

classify_character # For costMatrix
classify_characters # For cladisticMatrix
}
\examples{

# Example of a Type I character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 0,
    character_type = "unordered"
  )
)

# Example of a Type II character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 1,
    character_type = "unordered"
  )
)

# Example of a Type III character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 3,
    character_type = "unordered"
  )
)

# Example of a Type IV character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 3,
    character_type = "ordered"
  )
)

# Example of a Type V character:
classify_costmatrix(
  costmatrix = convert_adjacency_matrix_to_costmatrix(
    adjacency_matrix = matrix(
      data = c(
        0, 1, 0, 0,
        1, 0, 1, 1,
        0, 1, 0, 0,
        0, 1, 0, 0
      ),
      nrow = 4,
      dimnames = list(0:3, 0:3)
    )
  )
)

# Example of a Type VI character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 1,
    character_type = "stratigraphy",
    state_ages = c(103.7, 99.6)
  )
)

# Example of a Type VII character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 3,
    character_type = "irreversible"
  )
)

# Example of a Type VIII character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 1,
    character_type = "dollo"
  )
)

# Example of a Type IX character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 3,
    character_type = "dollo"
  )
)

# Example of a Type X character:

# Example of a Type XI character:

# Example of a Type XII character:
classify_costmatrix(
  costmatrix = make_costmatrix(
    min_state = 0,
    max_state= 3,
    character_type = "stratigraphy",
    state_ages = c(103.7, 99.6, 91.0, 77.2)
  )
)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
