% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompactifyMatrix.R
\name{CompactifyMatrix}
\alias{CompactifyMatrix}
\title{Collapses matrix to unique character state distributions}
\usage{
CompactifyMatrix(clad.matrix)
}
\arguments{
\item{clad.matrix}{The cladistic matrix in the format imported by \link{ReadMorphNexus}.}
}
\description{
Collapses matrix to unique character state distributions.
}
\details{
Removing characters or taxa from a matrix imported using \link{ReadMorphNexus} is not simple due to asscoiated vectors for ordering, character weights etc. To save repetitively pruning each part this function takes the matrix as input and vectors of either taxon names, character numbers, or one of each and removes those from the matrix. Minimum and maximum values (used by \link{MorphDistMatrix}) are also updated.
}
\examples{

# Remove the outgroup taxon and characters 11 and 53 from Gauthier1986:
prunedmatrix <- MatrixPruner(clad.matrix = Gauthier1986, taxa2prune = c("Outgroup"),
  characters2prune = c(11, 53))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\seealso{
\link{SafeTaxonomicReduction} and \link{MatrixPruner}
}
\keyword{NEXUS}

