% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Claddis-package.R
\docType{data}
\name{Day2016}
\alias{Day2016}
\title{Character-taxon matrix from Day et al. 2016}
\format{A character-taxon matrix in the format imported by \link{ReadMorphNexus}.}
\description{
The character-taxon matrix from Day et al. (2016).
}
\references{
Day, M. O., Rubidge, B. S. and Abdala, F., 2016. A new mid-Permian
burnetiamorph therapsid from the Main Karoo Basin of South Africa and a
phylogenetic review of Burnetiamorpha. Acta Palaeontologica Polonica, 61, 701-719.
}
\keyword{datasets}
