% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Claddis-package.R
\docType{data}
\name{day_2016}
\alias{day_2016}
\title{Character-taxon matrix from Day et al. 2016}
\format{
A character-taxon matrix in the format imported by \link{read_nexus_matrix}.
}
\description{
The character-taxon matrix from Day et al. (2016).
}
\references{
Day, M. O., Rubidge, B. S. and Abdala, F., 2016. A new mid-Permian burnetiamorph therapsid from the Main Karoo Basin of South Africa and a phylogenetic review of Burnetiamorpha. \emph{Acta Palaeontologica Polonica}, \bold{61}, 701-719.
}
\keyword{datasets}
