% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalize_weights.r
\name{equalize_weights}
\alias{equalize_weights}
\title{XXX}
\usage{
equalize_weights(cladistic_matrix, method)
}
\arguments{
\item{N}{The number of labels required,}
}
\value{
XXX.
}
\description{
XXX.
}
\details{
XXX.
}
\examples{

# XXX:
equalize_weights()

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
