% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ancestral_states.R
\name{estimate_ancestral_states}
\alias{estimate_ancestral_states}
\title{Ancestral Character State Estimation}
\usage{
estimate_ancestral_states(
  cladistic_matrix,
  time_tree,
  estimate.all.nodes = FALSE,
  estimate.tip.values = FALSE,
  inapplicables.as.missing = FALSE,
  polymorphism.behaviour = "equalp",
  uncertainty.behaviour = "equalp",
  threshold = 0.01,
  allow.all.missing = FALSE
)
}
\arguments{
\item{cladistic_matrix}{A character-taxon matrix in the format imported by \link{read_nexus_matrix}.}

\item{time_tree}{A tree (phylo object) with branch lengths that represents the relationships of the taxa in \code{cladistic_matrix}.}

\item{estimate.all.nodes}{Logical that allows the user to make estimates for all ancestral values. The default (\code{FALSE}) will only make estimates for nodes that link coded terminals (recommended).}

\item{estimate.tip.values}{Logical that allows the user to make estimates for tip values. The default (\code{FALSE}) will only makes estimates for internal nodes (recommended).}

\item{inapplicables.as.missing}{Logical that decides whether or not to treat inapplicables as missing (TRUE) or not (FALSE, the default and recommended option).}

\item{polymorphism.behaviour}{One of either "equalp" or "treatasmissing".}

\item{uncertainty.behaviour}{One of either "equalp" or "treatasmissing".}

\item{threshold}{The threshold value to use when collapsing marginal likelihoods to discrete state(s).}

\item{allow.all.missing}{Logical to allow all missing character values (generally not recommended, hence default is FALSE).}
}
\value{
\item{anc.lik.matrix}{A matrix of nodes (hypothetical ancestors; rows) against characters (columns) listing the reconstructed ancestral states.}
}
\description{
Given a tree and a cladistic matrix uses likelihood to estimate the ancestral states for every character.
}
\details{
At its' core the function uses either the \link{rerootingMethod} (Yang et al. 1995) as implemented in the \link{phytools} package (for discrete characters) or the \link{ace} function in the \link{ape} package (for continuous characters) to make ancestral state estimates. For discrete characters these are collapsed to the most likely state (or states, given equal likelihoods or likelihood within a defined threshold value). In the latter case the resulting states are represented as an uncertainty (i.e., states separated by a slash, e.g., 0/1. This is the method used by Brusatte et al. (2014).

The method can deal with ordered or unordered characters and does so by allowing only indirect transitions (from 0 to 2 must pass through 1) or direct transitions (from 0 straight to 2), respectively. However, more complex step matrix transitions are not currently supported.

Ancestral state estimation is complicated where polymorphic or
}
\examples{

# Set random seed:
set.seed(4)

# Generate a random tree for the Day data set:
time_tree <- rtree(n = nrow(day_2016$matrix_1$matrix))

# Update taxon names to match those in the data matrix:
time_tree$tip.label <- rownames(day_2016$matrix_1$matrix)

# Set root time by making youngest taxon extant:
time_tree$root.time <- max(diag(vcv(time_tree)))

# Use Day matrix as cladistic matrix:
cladistic_matrix <- day_2016

# Prune most characters out to make example run fast:
cladistic_matrix <- prune_cladistic_matrix(cladistic_matrix,
  characters2prune = c(2:3, 5:37))

# Estimate ancestral states:
estimate_ancestral_states(cladistic_matrix = cladistic_matrix,
  time_tree = time_tree)

}
\references{
Brusatte, S. L., Lloyd, G. T., Wang, S. C. and Norell, M. A., 2014. Gradual assembly of avian body plan culminated in rapid rates of evolution across dinosaur-bird transition. Current Biology, 24, 2386-2392.

Yang, Z., Kumar, S. and Nei, M., 1995. A new method of inference of ancestral nucleotide and amino acid sequences. Genetics, 141, 1641-1650.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com} and Thomas Guillerme \email{guillert@tcd.ie}
}
