% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_shortest_costmatrix_path.r
\name{find_shortest_costmatrix_path}
\alias{find_shortest_costmatrix_path}
\title{Finds the shortest path between two states in a costmatrix}
\usage{
find_shortest_costmatrix_path(costmatrix, start, end)
}
\arguments{
\item{costmatrix}{An object of class \code{costMatrix}.}

\item{start}{The start state for the requested path.}

\item{end}{The end state of the requested path.}
}
\value{
A vector of states describing (one) of the optimal path(s) in the order \code{start} to \code{end}.
}
\description{
Given a start and end state, returns the shortest path through a costmatrix.
}
\details{

}
\examples{

# Make a four-state Dollo costmatrix:
costmatrix <- make_costmatrix(
  min_state = 0,
  max_state = 3,
  character_type = "dollo",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle",
  dollo_penalty = 100
)

# Find the shortest path from state 0 to state 3:
find_shortest_costmatrix_path(
  costmatrix = costmatrix,
  start = "0",
  end = "3"
)

# Show that this is directional by asking for the reverse path:
find_shortest_costmatrix_path(
  costmatrix = costmatrix,
  start = "3",
  end = "0"
)

}
\references{
Dijkstra, E. W., 1959. A note on two problems in connexion with graphs. \emph{Numerische Mathematik}, \bold{1}, 269–271.

Hoyal Cuthill, J. F., Braddy, S. J. and Donoghue, P. C. J., 2010. A formula for maximum possible steps in multistate characters: isolating matrix parameter effects on measures of evolutionary convergence. \emph{Cladistics}, \bold{26}, 98-102.

Maddison, D. R. and Maddison, W. P., 2003. \emph{MacClade 4: Analysis of phylogeny and character evolution}. Version 4.06. Sinauer Associates, Sunderland, Massachusetts.
}
\seealso{
\link{convert_adjacency_matrix_to_costmatrix}, \link{make_costmatrix}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
