% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_steiner_tree_of_stategraph.R
\name{find_steiner_tree_of_stategraph}
\alias{find_steiner_tree_of_stategraph}
\title{Get edges of minimum spanning tree}
\usage{
find_steiner_tree_of_stategraph(stategraph, sampled_vertices)
}
\arguments{
\item{distance_matrix}{A square matrix of distances between objects.}
}
\value{
A vector of named edges (X->Y) with their distances. The sum of this vector is the length of the minimum spanning tree.
}
\description{
Returns edges of a minimum spanning tree given a distance matrix.
}
\details{
This function is a wrapper for \link{mst} in the \link{ape} package, but returns a vector of edges rather than a square matrix of links.
}
\examples{

# XXXX:

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
