% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stepmatrix_minimum_span.r
\name{find_stepmatrix_minimum_span}
\alias{find_stepmatrix_minimum_span}
\title{Finds the minimum spanning tree length for a stepmatrix}
\usage{
find_stepmatrix_minimum_span(stepmatrix)
}
\arguments{
\item{stepmatrix}{An object of class \code{stepMatrix}.}
}
\value{
A scalar indicating the ength in steps of the minimum spanning tree or arboresence.
}
\description{
Given a stepmatrix, returns the length of the shortest tree that connects every state.
}
\details{
The minimum parsimony length a phylogenetic hypothesis could have depends on the stepmatrix of transition costs and the states actually sampled. If the stepmatrix rows and columns already represent sampled states then this minimum length is reduced to a graph theory problem - the minimum spanning tree. This function returns the length of that tree and hence the minimum steps value (as used in stratocladistics to define "parsimony debt" Fisher 1994, or in homoplasy indices, see Hoyal Cuthill 2015).

Strictly speaking the answer depends on whether the stepmatrix is symmetric (representing an undirected graph) or asymmetric (representing a directed graph). In both cases the transition costs are formally the edge weights. If the former then the problem is the "regular" minimum spanning tree, and if the latter then it is known as the arboresence. If the former, then Kruskal's (1956) algorithm is used and if the latter then Edmonds' (1967) algorithm is used. (The function automatically makes this choice using the value of \code{stepmatrix$symmetry}. In practice the implementations used come from the optrees package (Fontenla 2014), and call the \code{getMinimumSpanningTree} or \code{msArborEdmonds} function, respectively.

Note that this function only returns the length of the answer (in steps) not the actual path.
}
\examples{

# Make a six-state ordered character stepmatrix:
stepmatrix <- make_stepmatrix(
  min_state = "0",
  max_state = "5",
  character_type = "ordered",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Find length of shortest spanning tree of stepmatrix:
find_stepmatrix_minimum_span(stepmatrix = stepmatrix)

}
\references{
Edmonds, J., 1967. Optimum branchings. \emph{Journal of Research of the National Bureau of Standards Section B}, \bold{71B}, 233-240.

Fisher, D. C., 1994. Stratocladistics: morphological and temporal patterns and their relation to phylogenetic process. In L. Grande and O. Rieppel (eds.), \emph{Interpreting the Hierarchy of Nature}. Academic Press, San Diego. pp133–171.

Fontenla, M., 2014. optrees: Optimal Trees in Weighted Graphs. R package version 1.0. https://CRAN.R-project.org/package=optrees

Hoyal Cuthill, J., 2015. The size of the character state space affects the occurrence and detection of homoplasy: modelling the probability of incompatibility for unordered phylogenetic characters. \emph{Journal ofTheoretical Biology}, \bold{366}, 24-32.

Kruskal, J. B., 1956. On the shortest spanning subtree of a graph and the traveling salesman problem. \emph{Proceedings of the American Mathematical Society}, \bold{7}, 48-50.
}
\seealso{
\link{find_shortest_path}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
