% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Claddis-package.R
\docType{data}
\name{Gauthier1986}
\alias{Gauthier1986}
\title{Character-taxon matrix from Gauthier 1986}
\format{A character-taxon matrix in the format imported by \link{ReadMorphNexus}.}
\description{
The character-taxon matrix from Gauthier (1986).
}
\references{
Gauthier, J. A., 1986. Saurischian monophyly and the origin of
birds. In Padian, K. (ed.) The Origin of Birds and the Evolution of Flight.
Towne and Bacon, San Francisco, CA, United States, 1-55.
}
\keyword{datasets}
