% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Claddis-package.R
\docType{data}
\name{gauthier_1986}
\alias{gauthier_1986}
\title{Character-taxon matrix from Gauthier 1986}
\format{
A character-taxon matrix in the format imported by \link{read_nexus_matrix}.
}
\description{
The character-taxon matrix from Gauthier (1986).
}
\references{
Gauthier, J. A., 1986. Saurischian monophyly and the origin of birds. In Padian, K. (ed.) \emph{The Origin of Birds and the Evolution of Flight}. Towne and Bacon, San Francisco, CA, United States, 1-55.
}
\keyword{datasets}
