% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.cladisticMatrix.R
\name{is.cladisticMatrix}
\alias{is.cladisticMatrix}
\title{Cladistic matrix class}
\usage{
is.cladisticMatrix(x)
}
\arguments{
\item{x}{A cladisticMatrix object.}
}
\value{
\code{is.cladisticMatrix} returns either TRUE or FALSE.
}
\description{
Functions to deal with the cladistic matrix class.
}
\details{
Claddis uses various classes to define specific types of data, here the use of cladistic data (the main input of the package) is assigned the class "cladisticMatrix".

\code{is.cladisticMatrix} checks whether an object is or is not a valid cladisticMatrix object.
}
\examples{

# Check that this is a valid cladisticMatrix object (will succeed as format and
# class are correct):
is.cladisticMatrix(x = day_2016)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
