% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.costMatrix.R
\name{is.costMatrix}
\alias{is.costMatrix}
\title{Costmatrix class}
\usage{
is.costMatrix(x)
}
\arguments{
\item{x}{An object of class \code{costMatrix}.}
}
\value{
\code{is.costMatrix} returns either TRUE or FALSE.
}
\description{
Functions to deal with the costmatrix class.
}
\details{
Claddis uses various classes to define specific types of data, here the use of costmatrices (to specify the parsimony costs of transitions between character states) are assigned the class "costMatrix".

\code{is.costMatrix} checks whether an object is or is not a valid costMatrix object.
}
\examples{

# Make an unordered costmatrix:
costmatrix <- make_costmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered"
)

# Check that this is a valid costMatrix object:
is.costMatrix(x = costmatrix)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
