% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.stepMatrix.R
\name{is.stepMatrix}
\alias{is.stepMatrix}
\title{Stepmatrix class}
\usage{
is.stepMatrix(x)
}
\arguments{
\item{x}{An object of class \code{stepMatrix}.}
}
\value{
\code{is.stepMatrix} returns either TRUE or FALSE.
}
\description{
Functions to deal with the stepmatrix class.
}
\details{
Claddis uses various classes to define specific types of data, here the use of stepmatrices (to specify the parsimony costs of transitions between character states) are assigned the class "stepMatrix".

\code{is.stepMatrix} checks whether an object is or is not a valid stepMatrix object.
}
\examples{

# Make an unordered stepmatrix:
stepmatrix <- make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Check that this is a valid stepMatrix object:
is.stepMatrix(x = stepmatrix)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
