% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_graph_connected.R
\name{is_graph_connected}
\alias{is_graph_connected}
\title{Is a graph connected?}
\usage{
is_graph_connected(laplacian_matrix)
}
\arguments{
\item{laplacian_matrix}{A Laplacian matrix (where the diagonal is the degree matrix and the off-diagonal the inverse of the adjacency matrix).}
}
\value{
A logical (TRUE or FALSE).
}
\description{
Is a graph connected?
}
\details{
Example graph:

\preformatted{6---4---5
    |   |\
    |   | 1
    |   |/
    3---2}

Has the adjacency matrix:

\preformatted{    _________________________
    | 1 | 2 | 3 | 4 | 5 | 6 |
-----------------------------
| 1 | 0 | 1 | 0 | 0 | 1 | 0 |
-----------------------------
| 2 | 1 | 0 | 1 | 0 | 1 | 0 |
-----------------------------
| 3 | 0 | 1 | 0 | 1 | 0 | 0 |
-----------------------------
| 4 | 0 | 0 | 1 | 0 | 1 | 1 |
-----------------------------
| 5 | 1 | 1 | 0 | 1 | 0 | 0 |
-----------------------------
| 6 | 0 | 0 | 0 | 1 | 0 | 0 |
-----------------------------}

And the degree matrix:

\preformatted{    _________________________
    | 1 | 2 | 3 | 4 | 5 | 6 |
-----------------------------
| 1 | 2 | 0 | 0 | 0 | 0 | 0 |
-----------------------------
| 2 | 0 | 3 | 0 | 0 | 0 | 0 |
-----------------------------
| 3 | 0 | 0 | 2 | 0 | 0 | 0 |
-----------------------------
| 4 | 0 | 0 | 0 | 3 | 0 | 0 |
-----------------------------
| 5 | 0 | 0 | 0 | 0 | 3 | 0 |
-----------------------------
| 6 | 0 | 0 | 0 | 0 | 0 | 1 |
-----------------------------}

And the Laplacian matrix:

\preformatted{    _________________________
    | 1 | 2 | 3 | 4 | 5 | 6 |
-----------------------------
| 1 | 2 |-1 | 0 | 0 |-1 | 0 |
-----------------------------
| 2 |-1 | 3 |-1 | 0 |-1 | 0 |
-----------------------------
| 3 | 0 |-1 | 2 |-1 | 0 | 0 |
-----------------------------
| 4 | 0 | 0 |-1 | 3 |-1 |-1 |
-----------------------------
| 5 |-1 |-1 | 0 |-1 | 3 | 0 |
-----------------------------
| 6 | 0 | 0 | 0 |-1 | 0 | 1 |
-----------------------------}
}
\examples{

x <- matrix(
  data = c(
     2, -1,  0,  0, -1,  0,
    -1,  3, -1,  0, -1,  0,
     0, -1,  2, -1,  0,  0,
     0,  0, -1,  3, -1, -1,
    -1, -1,  0, -1,  3,  0,
     0,  0,  0, -1,  0,  1
  ),
  ncol = 6,
  byrow = TRUE
)

# Check graph is connected:
is_graph_connected(laplacian_matrix = x)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
