% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_polymorphisms.r
\name{make_all_polymorphisms}
\alias{make_all_polymorphisms}
\title{Make all possible polymorphisms for a given set of states}
\usage{
make_all_polymorphisms(single_states)
}
\arguments{
\item{single_states}{A vector of single states (e.g., 0, 1, 2 etc.).}
}
\value{
A vector of all possible polymorphic (and non-polymorphic) states.
}
\description{
Given a set of discrete states will generate all possible polymorphic combinations of those states.
}
\details{
This function solves a simple phylogenetic combinatorics problem - what are all the possible outcomes for a character to be in given polymorphisms (of any size) are allowed?

For example, for three states (0, 1, 2) there are seven possible states inclusive of polymorphisms: 0, 1, 2, 0&1, 0&2, 1&2 and 0&1&2.

If the user is instead only interested in the size of this state space, this is simply given by 2^N - 1, where N is the number of single states. Thus, the first several outcomes are:

\preformatted{----------------------------------
| N states | N possible outcomes |
----------------------------------
|     2    |          3          |
|     3    |          7          |
|     4    |          15         |
|     5    |          31         |
|     6    |          63         |
|     7    |          127        |
|     8    |          255        |
|     9    |          511        |
|    10    |          1023       |
|    11    |          2047       |
|    12    |          4095       |
|    13    |          8191       |
|    14    |          16383      |
----------------------------------}

Note that this function is really designed for internal use, but may have value to some users and so is available "visibly" here.
}
\examples{

# Get all possible states for the character 0, 1, and 2:
make_all_polymorphisms(single_states = 0:2)

}
\seealso{
\link{make_costmatrix} and \link{make_all_uncertainties}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
