% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stepmatrix.r
\name{make_stepmatrix}
\alias{make_stepmatrix}
\title{Make a stepmatrix for a given set of states}
\usage{
make_stepmatrix(
  min_state = 0,
  max_state,
  character_type,
  include_polymorphisms = FALSE,
  polymorphism_shape,
  polymorphism_distance,
  state_ages,
  dollo_penalty = 999
)
}
\arguments{
\item{min_state}{The minimum character state (defaults to \code{0}).}

\item{max_state}{The maximum character state. Must be \code{1} or greater.}

\item{character_type}{The type of character desired. Must be one of: \code{"ordered"}, \code{"unordered"}, \code{"dollo"}, \code{"irreversible"}, or \code{"stratigraphy"}.}

\item{include_polymorphisms}{Logical indicating whether or not to include polymorphic state combinations (defaults to \code{FALSE}).}

\item{polymorphism_shape}{The shape to use for assigning polymorphism coordinates. Must be one of: \code{"hypercube"}, \code{"hypersphere"}, or \code{"simplex"}. (Only relevant if \code{character_type = "unordered"} and \code{include_polymorphisms = TRUE}.)}

\item{polymorphism_distance}{The distance to use to set transformation costs between polymorphic states. Must be one of: \code{"euclidean"}, \code{"great_circle"}, or \code{"manhattan"}.}

\item{state_ages}{A vector of ages assigned to each state (only relevant if \code{character_type = stratigraphy}).}

\item{dollo_penalty}{The size of the cost penalty for the acquisition of a Dollo character (defaults to \code{999}). Note: this should always be a positive real value greater than one, and never infinity (\code{Inf}), as at least one acquisition is expected.}
}
\value{
A square stepmatrix with rows representing "from" states, columns "to" states, and individual cells the cost in steps of that transition.
}
\description{
Given a set of discrete states and a character type will make the approriate stepmatrix for them.
}
\details{
Text.

\bold{Character types}
\emph{ordered} Wagner States must be in order of transistion, e.g., 0, 1, 2 means 0<->1<->2
\emph{unordered} Fitch
\emph{dollo} States must be in order.
\emph{irreversible} States must be in order. Aka. Camin-Sokal.
\emph{stratigraphy} States must be in order from oldest to youngest. Special case of irreversible and needs additional input of ages for differences!

\bold{Polymorphic characters}

Depends on character type too. E.g., for stratigraphy polymorphisms uses oldest state - leads to scenario where off-diagonal values can also be zero which may be important if looking for most parsimonious nacestral state reconstructions later. Not sure about Dollo and Irreversible?

\emph{Shape of polymorphic state space}

"hypercube", "hypersphere", "simplex"

\emph{Distances in polymorphic state space}

"manhattan", "euclidean", "great_circle" (assumes radius of one as does coordinates of shape)

\bold{Character weights}

Tempting to multiply space by weight to get weighted form, but no need as dealt with in parsimony function and can be dangerous with some values (e.g., weights of zero - may still want relative weights for ancestral state estimation even if not "charging" cost of transitions). Tranistions are mostly recslaed to a weight of one (except stratigraphy?).

\bold{More complex stepmatrices}

Can customise by mkaing one here as a starting point then manually modifying it. [CITE HOOKER]
}
\examples{

# Make an unordered stepmatrix:
make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Make an unordered stepmatrix including polymorphisms:
make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered",
  include_polymorphisms = TRUE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Make an ordered stepmatrix:
make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "ordered",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Make a Dollo stepmatrix:
make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "dollo",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle",
  dollo_penalty = 100
)

# Make an irreversible stepmatrix:
make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "irreversible",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Make a stratigraphic stepmatrix:
make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "stratigraphy",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle",
  state_ages = c(52, 34, 12)
)

# Make a stratigraphic stepmatrix with polymorphisms:
make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "stratigraphy",
  include_polymorphisms = TRUE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle",
  state_ages = c(52, 34, 12)
)

}
\references{
REF
}
\seealso{
\link{make_all_polymorphisms}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
