% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMorphMatrix.R
\name{MakeMorphMatrix}
\alias{MakeMorphMatrix}
\title{Creates a morphological data file from a matrix}
\usage{
MakeMorphMatrix(CTmatrix, header = "", weights = NULL, ordering = NULL,
  symbols = NULL, equalise.weights = FALSE)
}
\arguments{
\item{CTmatrix}{A Character-Taxon (columns-rows) matrix, with taxa names as rownames.}

\item{header}{A scalar indicating any header text (defaults to an empty string: "").}

\item{weights}{A vector specifying the weights used (if not specified defaults to 1).}

\item{ordering}{A vector indicating whether characters are ordered ("ord") or unordered ("unord") (if no specified defaults to ordered).}

\item{symbols}{The symbols to use if writing to a file (defaults to the numbers 0:9 then the letters A to V).}

\item{equalise.weights}{Optional that overrides the weights specified above make all characters truly equally weighted.}
}
\value{
\item{header}{Any header text included in the file given between square brackets as character.}
\item{matrix}{A matrix of taxa (rows) and characters (columns). The matrix is in character format in order to deal with polymorphisms, which are separated by ampersands.}
\item{ordering}{A character vector of the same length as the number of morphological characters indicating whether they are ordered (\code{ord}) or unordered (\code{unord}).}
\item{weights}{A numeric vector of the same length as the number of morphological characters indicating their weights.}
\item{max.vals}{A numeric vector of the same length as the number of morphological characters indicating the maximum state values.}
\item{min.vals}{A numeric vector of the same length as the number of morphological characters indicating the minimum state values.}
\item{step.matrices}{A list of any step matrices supplied in the input file. Is \code{NULL} if none are specified.}
\item{symbols}{The original symbols used in the input data (these are replaced in the matrix by integers starting at zero.}
}
\description{
Creates a morphological data file from a character-taxon matrix.
}
\details{
Claddis generally assumes that matrices will be imported into R from the #NEXUS format, but in some cases (e.g., when using simulated data) it might be desirable to build a matrix within R. This function allows the user to convert such a matrix into the format required by other Claddis functions.

NB: Currently the function cannot deal with step matrices or continuous characters.
}
\examples{

# Create random 10-by-50 matrix:
CTmatrix <- matrix(sample(c("0", "1", "0&1", NA), 500,
  replace = TRUE), nrow = 10, dimnames =
  list(apply(matrix(sample(LETTERS, 40,
  replace = TRUE), nrow = 10), 1, paste,
  collapse = ""), c()))

# Reformat for use elsewhere in Claddis:
MakeMorphMatrix(CTmatrix)

}
\seealso{
\link{ReadMorphNexus}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
