% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMorphNexus.R
\name{MakeMorphNexus}
\alias{MakeMorphNexus}
\title{Ancestral State Estimation}
\usage{
MakeMorphNexus(matrix, header, ordering, weights, combine.morph.matrix)
}
\arguments{
\item{matrix}{A discrete morphological characters matrix with rows as OTUs and columns as characters.}

\item{header}{An optional header text usually included in a \code{#NEXUS} file given between square brackets as character. If no provided, the header is set to be \code{""} (empty quotes).}

\item{ordering}{An optional vector of \code{"unord"} and \code{"ord"} character strings of the same length as the number of columns in the \code{matrix} argument. If not provided, all characters are assumed to be \code{"unord"}.}

\item{weights}{An optional of \code{numeric} values of the same length as the number of columns in the \code{matrix} argument. If not provided, all characters are assumed to be \code{1}.}

\item{combine.morph.matrix}{An optional character-taxon matrix in the format imported by \link{ReadMorphNexus} to combine with the \code{matrix} argument.}

\item{step.matrices}{An optional list of step matrices (default = \code{NULL}).}
}
\value{
\item{header}{Any header text included in the file given between square brackets as character.}
\item{matrix}{A matrix of taxa (rows) and characters (columns). The matrix is in character format in order to deal with polymorphisms, which are separated by ampersands.}
\item{ordering}{A character vector of the same length as the number of morphological characters indicating whether they are ordered (\code{ORD}) or unordered (\code{UNORD}).}
\item{weights}{A numeric vector of the same length as the number of morphological characters indicating their weights.}
\item{max.vals}{A numeric vector of the same length as the number of morphological characters indicating the maximum state values.}
\item{min.vals}{A numeric vector of the same length as the number of morphological characters indicating the minimum state values.}
\item{step.matrices}{A list of any step matrices supplied in the input file. Is \code{NULL} if none are specified.}
\item{symbols}{The original symbols used in the input data (these are replaced here by the digits 0-9 and the letters A-V in order).}
}
\description{
Transforms a single matrix into a list typically generated by the \link{ReadMorphNexus} function.
}
\details{
The \code{max.vals} and \code{min.vals} elements are automatically calculated from the input \code{matrix} argument.
}
\examples{

## Estimating ancestral states for the nodes of a matrix
# Set random seed:
set.seed(17)

# Generate a random tree for the Michaux data set:
tree <- rtree(nrow(Michaux1989$matrix))

# Update taxon names to match those in the data matrix:
tree$tip.label <- rownames(Michaux1989$matrix)

# Set root time by making youngest taxon extant:
tree$root.time <- max(diag(vcv(tree)))

# Estimate ancestral states:
matrix_nodes_estimates <- AncStateEstMatrix(Michaux1989, tree)

## Transform the matrix of ancestral nodes states into a ReadMorphNexus list
MakeMorphNexus(matrix_nodes_estimates)

## Combine it with the original matrix (containing the tips)
MakeMorphNexus(matrix_nodes_estimates, combine.morph.matrix = Michaux1989)

}
\author{
Thomas Guillerme \email{guillert@tcd.ie}
}
\seealso{
\link{ReadMorphNexus}, \link{WriteMorphNexus}
}

