% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dollo_changes.R
\name{map_dollo_changes}
\alias{map_dollo_changes}
\title{Stochastic Character Map For Dollo Character}
\usage{
map_dollo_changes(time_tree, tip_states)
}
\arguments{
\item{time_tree}{A tree in phylo format with positive branch lengths and a value for \code{$root.time}.}

\item{tip_states}{A named vector of tip states (must be 0 or 1), where the names match \code{tree$tip.label}.}
}
\value{
\item{changes}{A matrix of all changes (gains and losses).}
\item{stochastic_character_map}{The stochastic character map.}
}
\description{
Given a tree with binary tip states produces a stochastic Dollo character map.
}
\details{
The non-ideal solution from Tarver et al. (2018) to the problem of generating a stochastic character map for a Dollo character (i.e., a single gain of the derived state, 1) with any number of losses (1 -> 0).

The function operates as follows:

1) Establishes the least inclusive clade exhibiting the derived state (1).
2) Assumes a single gain occurred with equal probability along the branch subtending this clade.
3) Prunes the inclusive clade to generate a subtree with a strong root prior of the derived state (1).
4) Calls \code{make.simmap} from the \code{phytools} package to generate a stochastic character map using a model where only losses are possible.
5) Outputs both the stochastic character map (time spent in each state on each branch) and a matrix of state changes.

NB: As the map is stochastic the answer will be different each time the function is run and multiple replicates are strongly advised in order to characterise this uncertainty.
}
\examples{

# Create a random 10-taxon tree:
time_tree <- ape::rtree(n = 10)

# Arbitrarily add a root.time value of 100 Ma:
time_tree$root.time <- 100

# Generate random tip states (0s and 1s):
tip_states <- sample(c(0, 1), 10, replace = TRUE)

# Add labels to tip states:
names(tip_states) <- time_tree$tip.label

# Get a single stochastic character map:
out <- map_dollo_changes(time_tree, tip_states)

# View matrix of changes:
out$changes

# View stochastic character map (time spent in each state on each branch):
out$scm
}
\references{
Tarver, J. E., Taylor, R. S., Puttick, M. N., Lloyd, G. T., Pett, W., Fromm, B., Schirrmeister, B. E., Pisani, D., Peterson, K. J. and Donoghue, P. C. J., 2018. Well-annotated microRNAomes do not evidence pervasive miRNA loss. \emph{Genome Biology and Evolution}, \bold{6}, 1457-1470.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
