% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stochastic_changes.R
\name{map_stochastic_changes}
\alias{map_stochastic_changes}
\title{Finds all state changes on a tree using stochastic character mapping}
\usage{
map_stochastic_changes(
  cladistic_matrix,
  time_tree,
  time_bins,
  NSimulations = 10,
  polymorphism.behaviour = "equalp",
  uncertainty.behaviour = "equalp",
  inapplicable.behaviour = "missing"
)
}
\arguments{
\item{cladistic_matrix}{A character-taxon matrix in the format imported by \link{read_nexus_matrix}.}

\item{time_tree}{A time-scaled tree (phylo object) that represents the relationships of the taxa in \code{cladistic_matrix}.}

\item{time_bins}{A vector of ages representing the boundaries of a series of time bins.}

\item{NSimulations}{The number of simulations to perform (passed to \code{make.simmap}.}

\item{polymorphism.behaviour}{What to do with polymorphic (&) characters. One of "equalp", "missing", or "random". See details.}

\item{uncertainty.behaviour}{What to do with uncertain (/) characters. One of "equalp", "missing", or "random". See details.}

\item{inapplicable.behaviour}{What to do with inapplicable characters. Only one option currently ("missing"). See details.}
}
\value{
\item{RootStates}{A matrix of the root states for each character (column) and simulation (rows).}
\item{AllStateChanges}{A matrix of rows for each change with columns corresponding to the character, the simulation number, the edge number, the time the change occurred, and the start and end states.}
\item{CharacterTimes}{A vector of the sampled tree-length (in Ma) for each character.}
\item{EdgeLengthsPerBin}{A matrix of time bins (columns) and characters (rows) indicating the sampled tree-length (in Ma).}
\item{TerminalEdgeLengthsPerBin}{As above, but for terminal edges only.}
\item{InternalEdgeLengthsPerBin}{As above, but for internal edges only.}
}
\description{
Takes a cladistic matrix and time-scaled tree and makes point estimates for every character change using stochastic character mapping.
}
\details{
Important: this function is not yet complete and should not be used.

A wrapper function for \link{make.simmap} in the \link{phytools} package.

This function is intended to enumerate all possible changes on a tree (including to and from missing or inapplicable states) under the assumptions of stochastic character mapping as an alternative means of establishing branch-lengths (for rate analyses) or recording the state occupied at a particular point in time for disparity analyses.
}
\examples{

# Set random seed:
set.seed(2)

# Use Day 2016 as source matrix:
cladistic_matrix <- day_2016

# Prune out continuous characters:
cladistic_matrix <- prune_cladistic_matrix(cladistic_matrix =
  cladistic_matrix, blocks2prune = 1)

# Prune out majority of characters so
# example runs quickly:
cladistic_matrix <- prune_cladistic_matrix(cladistic_matrix =
  cladistic_matrix, characters2prune = 1:32)

# Generete random tree for matrix taxa:
time_tree <- rtree(nrow(day_2016$matrix_1$matrix))

# Add taxon names to tree:
time_tree$tip.label <- rownames(day_2016$matrix_1$matrix)

# Add root age to tree:
time_tree$root.time <- max(diag(vcv(time_tree)))

# Get all state changes for two simulations:
StateChanges <-
  map_stochastic_changes(cladistic_matrix = cladistic_matrix,
  time_tree = time_tree, time_bins = seq(time_tree$root.time, 0,
  length.out = 3), NSimulations = 2)

# View matrix of all stochstic character changes:
StateChanges$AllStateChanges

# View vector of sampled time for each
# character:
StateChanges$CharacterTimes

# View matrix of edge lengths in each time bin:
StateChanges$EdgeLengthsPerBin

# View matrix of termnial edge lengths in each time bin:
StateChanges$TerminalEdgeLengthsPerBin

# View matrix of internal edge lengths in each time bin:
StateChanges$InternalEdgeLengthsPerBin

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
