% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixPruner.R
\name{MatrixPruner}
\alias{MatrixPruner}
\title{Prunes a character matrix of characters or taxa}
\usage{
MatrixPruner(clad.matrix, taxa2prune = c(), characters2prune = c())
}
\arguments{
\item{clad.matrix}{The cladistic matrix in the format imported by \link{ReadMorphNexus}.}

\item{taxa2prune}{A vector of taxon names to prune (these must be present in \code{rownames(clad.matrix$matrix}).}

\item{characters2prune}{A vector of character numbers to prune.}
}
\description{
Prunes a character matrix of characters, taxa, or both.
}
\details{
Removing characters or taxa from a matrix imported using \link{ReadMorphNexus} is not simple due to associated vectors for ordering, character weights etc. To save repetitively pruning each part this function takes the matrix as input and vectors of either taxon names, character numbers, or both and removes those from the matrix. Minimum and maximum values (used by \link{MorphDistMatrix}) are also updated.
}
\examples{

# Remove the outgroup taxon and characters 11 and 53 from Gauthier1986:
prunedmatrix <- MatrixPruner(clad.matrix = Gauthier1986, taxa2prune = c("Outgroup"),
  characters2prune = c(11, 53))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\seealso{
\link{ReadMorphNexus}
}
\keyword{NEXUS}

