\name{Michaux1989}

\docType{data}

\alias{Michaux1989}

\title{Character-taxon matrix from Michaux 1989}

\description{The character-taxon matrix from Michaux (1989).}

\usage{data(Michaux1989)}

\format{A character-taxon matrix in the format imported by \link{ReadMorphNexus}.}

\references{
Michaux, B., 1989. Cladograms can reconstruct phylogenies: an example from the fossil record. Alcheringa, 13, 21-36.
}

\keyword{ datasets }
