% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Claddis-package.R
\docType{data}
\name{michaux_1989}
\alias{michaux_1989}
\title{Character-taxon matrix from Michaux 1989}
\format{
A character-taxon matrix in the format imported by
\link{read_nexus_matrix}.
}
\description{
The character-taxon matrix from Michaux (1989).
}
\references{
Michaux, B., 1989. Cladograms can reconstruct phylogenies: an example from the fossil record. \emph{Alcheringa}, \bold{13}, 21-36.
}
\keyword{datasets}
