% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorphDistMatrixFast.R
\name{MorphDistMatrixFast}
\alias{MorphDistMatrixFast}
\title{Get distance matrices from a cladistic matrix}
\usage{
MorphDistMatrixFast(morph.matrix, distance = c("Raw", "GED", "Gower", "Max",
  "Comp"), transform.proportional.distances = "arcsine_sqrt")
}
\arguments{
\item{morph.matrix}{A character-taxon matrix in the format imported by \link{ReadMorphNexus}.}

\item{distance}{Which distance matrices to return. Can be one or more of the following \code{"Raw"}, \code{"GED"}, \code{"Gower"}, \code{"Max"}, \code{"Comp"}. By default, the functions returns all 5 matrices.}

\item{transform.proportional.distances}{Whether to transform the proportional distances (Gower and max). Options are \code{none}, \code{sqrt}, or \code{arcsine_sqrt} (the default).}
}
\value{
\item{raw.dist.matrix}{A distance matrix indicating the raw distances (based on the method set in \code{dist.method}) of each pairwise comparison.}
\item{GED.dist.matrix}{A Generalised Euclidean Distance (GED) matrix (Wills 2001).}
\item{gower.dist.matrix}{A distance matrix where raw distances have been rescaled using Gower (1971) dissimilarity (then rescaled from 0 to 1 and the arcsine of the square root taken).}
\item{max.dist.matrix}{A distance matrix where raw distances have been rescaled against the maximum possible observable distance (then the arcsine of the square root taken).}
\item{comp.char.matrix}{A matrix showing the number of coded characters that were used to make each pairwise comparison.}
}
\description{
Takes a cladistic morphological dataset and converts it into a set of
pairwise distances.
}
\details{
Essentially the same function as \link{MorphDistMatrixFast}, but refactored for speed.
}
\examples{

# Get morphological distances for Michaux (1989)
# data set:
distances <- MorphDistMatrixFast(Michaux1989)

# Show raw Euclidean distances:
distances$raw.dist.matrix

# Show Generailsed Euclidean Distances:
distances$GED.dist.matrix

# Show Gower rescaled distances:
distances$gower.dist.matrix

# Show maximum observable rescaled distances:
distances$max.dist.matrix

# Show number of characters that can be scored for
# each pairwise comparison:
distances$comp.char.matrix

}
\author{
Thomas Guillerme \email{guillert@tcd.ie}
}
\keyword{distance}

