% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiMorphospacePlot.R
\name{MultiMorphospacePlot}
\alias{MultiMorphospacePlot}
\title{Plot Multiple Morphopaces}
\usage{
MultiMorphospacePlot(pcoa_input, N_axes = 4, plot_taxon_names = FALSE,
  plot_internal_nodes = FALSE, plot_root = TRUE, root_colour = "grey")
}
\arguments{
\item{pcoa_input}{Text.}

\item{N_axes}{Text.}

\item{plot_taxon_names}{Text.}

\item{plot_internal_nodes}{Text.}

\item{plot_root}{Text.}

\item{root_colour}{Text.}
}
\description{
Plots morphospaces for any number of axes using the output from MorphMatrix2PCoA.
}
\details{
Input must come from \link{MorphMatrix2PCoA}. Is a wrapper for \link{MorphospacePlot}.
}
\examples{
# Nothing yet
}
\author{
Emma Sherratt \email{emma.sherratt@gmail.com} and Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{coordinates}
\keyword{principal}

