% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_combinations_with_replacement.r
\name{permute_combinations_with_replacement}
\alias{permute_combinations_with_replacement}
\title{Permute all combinations of x of size m with replacement}
\usage{
permute_combinations_with_replacement(x, m)
}
\arguments{
\item{x}{A character vector.}

\item{m}{A positive integer indicating the size of the set desired.}
}
\value{
A matrix of m columns where each row is a unique combination of x.
}
\description{
Given a vector x, permutes all possible groups of size m ignoring order and allowing any item in x to appear multiple times.
}
\details{
This is a simple combinatoric function used internally in Claddis where all possible combinations of \code{x} that are size \code{m} are permuted. Note that this ignores order (i.e., the sets \\{A,B\\} and \\{B,A\\} are considered identical) and replacements (or multiples) of an element of \code{x} are allowed (i.e., the sets \\{A,A\\} and \\{B,B\\} are both valid).
}
\examples{

# Permute all the ways the letters A-C can form a set of size 3:
permute_combinations_with_replacement(x = LETTERS[1:3], m = 3)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
