% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_changes_on_tree.R
\name{plot_changes_on_tree}
\alias{plot_changes_on_tree}
\title{Plots character changes on branches}
\usage{
plot_changes_on_tree(character_changes, time_tree)
}
\arguments{
\item{character_changes}{A matrix of character changes.}

\item{time_tree}{Tree on which character changes occur.}
}
\value{
A plot of character changes on a tree.
}
\description{
Plots character changes in boxes on branches.
}
\details{
Takes the \code{character_changes} output from \link{test_rates} and plots it on the tree used to generate it.
}
\examples{

# Set random seed:
set.seed(17)

# Generate a random tree for the Michaux data set:
time_tree <- ape::rtree(nrow(michaux_1989$matrix_1$matrix))

# Update taxon names to match those in the data matrix:
time_tree$tip.label <- rownames(michaux_1989$matrix_1$matrix)

# Set root time by making youngest taxon extant:
time_tree$root.time <- max(diag(vcv(time_tree)))

# Get discrete character rates (includes changes):
out <- test_rates(time_tree, michaux_1989,
  seq(time_tree$root.time, 0, length.out = 3),
  branch_partitions = list(list(1)), alpha = 0.01
)

# Plot character changes on the tree:
plot_changes_on_tree(
  out$InferredCharacterChanges,
  time_tree
)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
