% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chronophylomorphospace.R
\name{plot_chronophylomorphospace}
\alias{plot_chronophylomorphospace}
\title{Chronophylomorphospace Plot}
\usage{
plot_chronophylomorphospace(
  pcoa_data,
  x_axis = 1,
  y_axis = 2,
  plot_tips = TRUE,
  plot_nodes = TRUE,
  show_names = TRUE,
  plot_branches = TRUE,
  shadow = TRUE
)
}
\arguments{
\item{pcoa_data}{Principal coordinate data in the format output by  \link{ordinate_cladistic_matrix} that includes a tree and ancestral states.}

\item{x_axis}{Which ordination axis to plot as the x-axis.}

\item{y_axis}{Which ordination axis to plot as the y-axis.}

\item{plot_tips}{Whether or not to plot the tip nodes (defaults to TRUE).}

\item{plot_nodes}{Whether or not to plot the internal nodes (defaults to TRUE).}

\item{show_names}{Whether or not to show the taxon nodes (defaults to TRUE).}

\item{plot_branches}{Whether or not to plot the branches (defaults to TRUE).}

\item{shadow}{Whether or not to plot a shadow (2D plot) on the bottom face of the 3D plot (defaults to TRUE).}
}
\description{
Plots a three-dimensional chronophylomorphospace.
}
\details{
Creates a manually repositionable three-dimensional (two ordination axes plus time) plot of a phylomorphospace.

This function aims to mimic the data visualisation of Sakamoto and Ruta (2012; their Video S1).
}
\examples{

\dontrun{
require(rgl)

# Set random seed:
set.seed(4)

# Generate a random tree for the Michaux 1989 data set:
time_tree <- ape::rtree(n = nrow(michaux_1989$matrix_1$matrix))

# Set root time so latest tip terminates at the present:
time_tree$root.time <- max(diag(x = ape::vcv(phy = time_tree)))

# Add taxon names to the tree:
time_tree$tip.label <- rownames(x = michaux_1989$matrix_1$matrix)

# Perform a phylogenetic Principal Coordinates Analysis:
pcoa_data <- ordinate_cladistic_matrix(michaux_1989,
  time_tree = time_tree
)

# Plot a chronophylomorphospace:
plot_chronophylomorphospace(pcoa_data)
}

}
\references{
Sakamoto, M. and Ruta, M. 2012. Convergence and divergence in the evolution of cat skulls: temporal and spatial patterns of morphological diversity. \emph{PLoS ONE}, \bold{7}, e39752.
}
\seealso{
\link{assign_taxa_to_bins}, \link{plot_morphospace_stack}, \link{plot_morphospace}, \link{plot_multi_morphospace}, \link{ordinate_cladistic_matrix}
}
\author{
Emma Sherratt \email{emma.sherratt@gmail.com} and Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
