% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_morphospace.R
\name{plot_morphospace}
\alias{plot_morphospace}
\title{Plot Morphopace}
\usage{
plot_morphospace(
  pcoa_input,
  x_axis = 1,
  y_axis = 2,
  z_axis = NULL,
  plot_taxon_names = FALSE,
  plot_internal_nodes = FALSE,
  plot_root = TRUE,
  root_colour = "grey"
)
}
\arguments{
\item{pcoa_input}{The main input in the format output from \link{ordinate_cladistic_matrix}.}

\item{x_axis}{Which ordination axis to plot as the x-axis (defaults to 1).}

\item{y_axis}{Which ordination axis to plot as the y-axis (defaults to 2).}

\item{z_axis}{Which ordination axis to plot as the z-axis (defaults to NULL, i.e., is not plotted).}

\item{plot_taxon_names}{Optional to plot the names of the taxa (defaults to FALSE).}

\item{plot_internal_nodes}{Optional to plot the internal nodes of the tree (if included in \code{pcoa_input}) (defaults to FALSE).}

\item{plot_root}{Optional to plot the root separately (defaults to FALSE).}

\item{root_colour}{If plotting the root separately (previous option) sets the root colour.}
}
\description{
Plots a morphospace using the output from ordinate_cladistic_matrix.
}
\details{
Uses output from \link{ordinate_cladistic_matrix} as input.

Allows plotting of a third axis using the technique of Matthew Wills (Wills et al. 1994; their Figures 4 and 8; Wills 1998; his Figure 4), where black and white indicate positive and negative values respectovely, and the size of points there magnitudes.
}
\examples{

# Set random seed:
set.seed(4)

# Generate a random tree for the Michaux 1989 data set:
time_tree <- ape::rtree(n = length(x = rownames(x = michaux_1989$matrix_1$matrix)))

# Add taxon names to the tree:
time_tree$tip.label <- rownames(x = michaux_1989$matrix_1$matrix)

# Perform a phylogenetic Principal Coordinates Analysis:
pcoa_input <- ordinate_cladistic_matrix(cladistic_matrix = michaux_1989, time_tree = time_tree)

# Plot the results:
plot_morphospace(pcoa_input = pcoa_input, plot_taxon_names = TRUE)
}
\references{
Wills, M. A., 1998. Cambrian and Recent disparity: the picture from priapulids. \emph{Paleobiology}, \bold{24}, 177-199.

Wills, M. A., Briggs, D. E. G. and Fortey, R. A., 1994. Disparity as an evolutionary index: a comparison of Cambrian and Recent arthropods. \emph{Paleobiology}, \bold{20}, 93-130.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com} and Emma Sherratt \email{emma.sherratt@gmail.com}
}
