% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rates_time.R
\name{plot_rates_time}
\alias{plot_rates_time}
\title{Visualize a rate test time series}
\usage{
plot_rates_time(test_rates_output, model_number, ...)
}
\arguments{
\item{test_rates_output}{Rate output from \link{test_rates}.}

\item{model_number}{The number of the model you wish to visualise from the rate output.}

\item{...}{Other options to be passed to \link{geoscalePlot}.}
}
\value{
Nothing is returned, but a plot is produced.
}
\description{
Given the results from a rates test produces a time series visualization for a specific model.
}
\details{
The raw output from \link{test_rates} can be difficult to interpret without visualization and this function provides a means for doing that when the desired output is a time series (other functions will be added for other types of rate test).

The function will only work for a single model, but in practice the user may wish to produce multiple plots in which case they simply need to rn the function multiple times or setup a multipanel window first with \link{layout}, or similar.

Plots use the \link{geoscale} package to add geologic time to the x-axis and interested users should consult the documentation tere for a full ist of options (passed via ...) in the function (see example below).

Calculated rates (changes per lineage million years) are plotted as filled circles and models are plotted as horizontal lines labelled by rate parameters (lambda_i).
}
\examples{

\dontrun{
Ages <- read.table("~/Documents/Packages/Claddis/LungfishTest/ages.txt",
  sep =",")
Matrix <-
  Claddis::read_nexus_matrix("~/Documents/Packages/Claddis/LungfishTest/Lloyd_etal_2012a.nex")
time_tree <-
  ape::read.tree("~/Documents/Packages/Claddis/LungfishTest/Lloyd_etal_2012a.tre")
time_tree <- time_tree[sample(1:100000, 100)]
time_tree <- lapply(time_tree, function(x) strap::DatePhylo(x, Ages, rlen = 2, method = "equal"))
class(time_tree) <- "multiPhylo"
time_bins <- c(443.8, 358.9, 298.9, 251.9, 201.3, 145.0, 66.0, 0.0)
LungfishResults <- Claddis::test_rates(time_tree[[1]], Matrix, time_bins,
  time_partitions = partition_time_bins(7),
  character_partitions = list(list(1:91), list(Cranial = 1:81,
  Postcranial = 82:91)))
plot_rates_time(LungfishResults,
  which(geiger::aicw(unlist(lapply(LungfishResults$TimeBinResults,
  function(x) x$AIC)))[, "delta"] == 0), units = "Period", cex.ts = 1,
  cex.age = 1, abbrev = "Period")
}

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
