% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCharacterChanges.R
\name{PlotCharacterChanges}
\alias{PlotCharacterChanges}
\title{Plots character changes on branches}
\usage{
PlotCharacterChanges(character.changes, tree)
}
\arguments{
\item{character.changes}{A matrix of character changes.}

\item{tree}{Tree on which character changes occur.}
}
\value{
A plot of character changes on a tree.
}
\description{
Plots character changes on branches.
}
\examples{

# Nothing yet

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

