% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cladisticMatrix.r
\name{print.cladisticMatrix}
\alias{print.cladisticMatrix}
\title{Compact display of a cladistic matrix}
\usage{
\method{print}{cladisticMatrix}(cladistic_matrix)
}
\arguments{
\item{cladistic_matrix}{An object of class \code{"cladisticMatrix"}.}
}
\value{
Nothing is directly retrned, instead a text summary describing the dimensions and nature of an object of class \code{"cladisticMatrix"} is printed to the console.
}
\description{
Displays a compact summary of the dimensions and nature of a cladistic matrix object.
}
\details{
Displays some basic summary information on a cladistic matrix object, including number and type of characters, information about ordering, and whether variable weights are used.
}
\examples{

# Show print.cladisticMatrix version of each included data sets:
print.cladisticMatrix(cladistic_matrix = day_2016)
print.cladisticMatrix(cladistic_matrix = gauthier_1986)
print.cladisticMatrix(cladistic_matrix = michaux_1989)

}
\seealso{
\link{build_cladistic_matrix}, \link{compactify_matrix}, \link{prune_cladistic_matrix}, \link{read_nexus_matrix}, \link{safe_taxonomic_reduction}, \link{write_nexus_matrix}, \link{write_tnt_matrix}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
