% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.costMatrix.r
\name{print.costMatrix}
\alias{print.costMatrix}
\title{Compact display of a costmatrix}
\usage{
\method{print}{costMatrix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"costMatrix"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Nothing is directly returned, instead a text summary describing a \code{"costMatrix"} object is printed to the console.
}
\description{
Displays a compact summary of a costMatrix object.
}
\details{
Displays some basic summary information on a costmatrix object.
}
\examples{

# Make an unordered costmatrix:
example_costmatrix <- make_costmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered"
)

# Show print.costMatrix version:
print.costMatrix(x = example_costmatrix)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
