% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stateGraph.r
\name{print.stateGraph}
\alias{print.stateGraph}
\title{Compact display of a stategraph}
\usage{
\method{print}{stateGraph}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"stateGraph"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Nothing is directly returned, instead a text summary describing a \code{"stateGraph"} object is printed to the console.
}
\description{
Displays a compact summary of a stateGraph object.
}
\details{
Displays some basic summary information on a stateGraph object.
}
\examples{

# Make an unordered costmatrix:
example_stategraph <- make_costmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered"
)

# Show print.stateGraph version:
print.stateGraph(x = example_stategraph)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
