% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stepMatrix.r
\name{print.stepMatrix}
\alias{print.stepMatrix}
\title{Compact display of a step matrix}
\usage{
\method{print}{stepMatrix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"stepMatrix"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Nothing is directly returned, instead a text summary describing a \code{"stepMatrix"} object is printed to the console.
}
\description{
Displays a compact summary of a stepMatrix object.
}
\details{
Displays some basic summary information on a stepmatrix object.
}
\examples{

# Make an unordered stepmatrix:
example_stepmatrix <- make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Show print.stepMatrix version of each included data sets:
print.stepMatrix(x = example_stepmatrix)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
