% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_homoplasy_indices.r
\name{prune_costmatrix}
\alias{prune_costmatrix}
\title{This is the type of graph where a state (0) - if unsampled - could not be pruned as it would lead to incorrect inferences. I.e., 0 is a cut vertex.
 1     2
  \   /
   \ /
    0
    |
    |
    3}
\usage{
prune_costmatrix(costmatrix, sampled_states, message = TRUE)
}
\description{
Costmatrix representation:
}
\details{
\preformatted{    -----------------
    | 0 | 1 | 2 | 3 |
---------------------
| 0 | 0 | 1 | 1 | 1 |
---------------------
| 1 | 1 | 0 | 2 | 2 |
---------------------
| 2 | 1 | 2 | 0 | 2 |
---------------------
| 3 | 1 | 2 | 2 | 0 |
---------------------}
}
