% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tree_length.r
\name{reconstruct_ancestral_states}
\alias{reconstruct_ancestral_states}
\title{Determine maximum parsimony ancestral state reconstruction(s)}
\usage{
reconstruct_ancestral_states(trees)
}
\arguments{
\item{trees}{A tree (phylo object) or set of trees (multiPhylo object).}
}
\value{
A list with multiple components, including:

\item{ITEM}{ITEM EXPLANATION.}
}
\description{
Given a tree, or set of trees, and a cladistic matrix returns all most parsimonious reconstruction(s) for every character.
}
\details{
Users may also wish to refer to the more complex whole-matrix, likelihood-based function \link{estimate_ancestral_states}. Although, note that eventually parsimony will simply be an option to that function.
}
\examples{

# TO DO

}
\references{
Swofford, D. L. and Maddison, W. P., 1992. Parsimony, character-state reconstructions, and evolutionary inferences. \emph{In} R. L. Mayden (ed.) Systematics, Historical Ecology, and North American Freshwater Fishes. Stanford University Press, Stanford, p187-223.
}
\seealso{
\link{estimate_ancestral_states}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
