% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_taxonomic_reinsertion.R
\name{safe_taxonomic_reinsertion}
\alias{safe_taxonomic_reinsertion}
\title{Reinsert Safely Removed Taxa Into A Tree}
\usage{
safe_taxonomic_reinsertion(
  input_filename,
  output_filename,
  str_taxa,
  multiple_placement_option = "exclude"
)
}
\arguments{
\item{input_filename}{A Newick-formatted tree file containing tree(s) without safely removed taxa.}

\item{output_filename}{A file name where the newly generated trees will be written out to (required).}

\item{str_taxa}{The safe taxonomic reduction table as generated by \link{safe_taxonomic_reduction}.}

\item{multiple_placement_option}{What to do with taxa that have more than one possible reinsertion position. Options are \code{"exclude"} (does not reinsert them; the default) or \code{"random"} (picks one of the possible positions and uses that - will vary stochastically if multiple trees exist).}
}
\value{
Nothing is returned, but a new file (\code{output_filename}) is written.
}
\description{
Safely reinsert taxa in a tree after they were removed from a matrix by Safe Taxonomic Reduction.
}
\details{
TEXT NEEDED!

OPERATES ON TEXT FILES, NOT APE IMPORTED TREES AS ASSUMES VERY LARGE FILES AND HENCE WILL BE SLOW IF RUN ON APE FORMAT TREES
}
\examples{

# Nothing yet
}
\seealso{
\link{safe_taxonomic_reduction}
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
