% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StackPlot.R
\name{StackPlot}
\alias{StackPlot}
\title{Plot Stacked Ordination Spaces}
\usage{
StackPlot(ordination_axes, ages, groups = NULL, time_slices, shear = 0.2,
  x_axis = 1, y_axis = 2)
}
\arguments{
\item{ordination_axes}{A matrix of the ordination axes supplied (rownames should be object names). First column should be values for first axis, second for second axis and so on.}

\item{ages}{A two-column matrix of the first and last apperance dates for the taxa in the same format supplied to \link{DatePhylo}.}

\item{groups}{Left as NULL for now.}

\item{time_slices}{A vector of the boundaries for a series of time slices.}

\item{shear}{A single value (0 to 1) for the degree of shearing in the stacked ordination spaces.}

\item{x_axis}{The ordination axis to plot on the x-axis.}

\item{y_axis}{The ordination axis to plot nn the y-axis.}
}
\description{
Plots a stack of ordination spaces representing multiple time-slices.
}
\details{
Default is oldest at bottom to youngest at top.
}
\examples{
# Nothing yet
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com} and Emma Sherratt \email{emma.sherratt@gmail.com}
}
\references{
Friedman coelacanth paper and others?
}
\keyword{coordinates}
\keyword{principal}

