% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_marginal_whitespace.r
\name{trim_marginal_whitespace}
\alias{trim_marginal_whitespace}
\title{Trims marginal whitespace}
\usage{
trim_marginal_whitespace(x)
}
\arguments{
\item{x}{A character string}
}
\value{
A vector of character string(s) with any leading or trailing whitespace removed.
}
\description{
Trims any marginal whitespace from a vector of character string(s).
}
\details{
Trims any marginal whitespace (spaces or tabs) from a vector of character string(s).
}
\examples{

# Example string:
x <- "   \td s f\t s  "

# Trim only marginal whitespace:
trim_marginal_whitespace(x)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
