% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stepMatrix.R
\name{check_stepMatrix}
\alias{check_stepMatrix}
\title{Check stepMatrix object for errors}
\usage{
check_stepMatrix(stepmatrix)
}
\arguments{
\item{stepmatrix}{A stepMatrix object.}
}
\value{
An error message or empty vector if no errors found.
}
\description{
Internal function to check stepMatrix object for errors.
}
\details{
Internal Claddis function. Nothing to see here. Carry on.

Solves issue raised in Maddison and Maddison (2003) by checking custom stepmatrices are internally consistent such that each transition cost represents the cost of the shortest possible path.
}
\examples{

# Make an unordered stepmatrix:
stepmatrix <- make_stepmatrix(
  min_state = 0,
  max_state = 2,
  character_type = "unordered",
  include_polymorphisms = FALSE,
  polymorphism_shape = "hypersphere",
  polymorphism_distance = "great_circle"
)

# Check that this is a valid stepMatrix object (should return empty vector):
check_stepMatrix(stepmatrix = stepmatrix)

}
\references{
Maddison, D. R. and Maddison, W. P., 2003. \emph{MacClade 4: Analysis of phylogeny and character evolution}. Version 4.06. Sinauer Associates, Sunderland, Massachusetts.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
