\name{Claddis-package}

\alias{Claddis-package}

\alias{Claddis}

\docType{package}

\title{
Cladistic disparity
}

\description{
A package to read cladistic morphological datasets into R and 
cpnvert them into distance matrices to enable multivariate 
ordinations to be performed upon them or perform analyses of
rates of discrete character evolution.
}

\details{
\tabular{ll}{
Package: \tab Claddis\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-05-22\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Graeme T. Lloyd <graemetlloyd@gmail.com>
}

\references{
Brusatte et al. 2008
Brusatte et al. 2014
Lloyd et al. 2012
}

\keyword{ disparity,distance,morphology,phylogeny }

\examples{
# Nothing yet
}