% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Claddis-package.R
\docType{package}
\name{Claddis-package}
\alias{Claddis}
\alias{Claddis-package}
\title{Measuring Morphological Diversity and Evolutionary Tempo}
\description{
Measures morphological diversity from discrete character data and estimates evolutionary tempo on phylogenetic trees.
}
\examples{
# Get morphological distances for Michaux (1989) data set:
distances <- MorphDistMatrix(Michaux1989)

# Show distances:
distances
}
\author{
Graeme T. Lloyd <graemetlloyd@gmail.com>
}
\references{
Brusatte, S. L., Benton, M. J., Ruta, M. and Lloyd, G. T.,
2008a. Superiority, competition, and opportunism in the evolutionary
radiation of dinosaurs. Science, 321, 1485-1488.

Brusatte, S. L., Benton, M. J., Ruta, M. and Lloyd, G. T., 2008b. The first
50 million years of dinosaur evolution: macroevolutionary pattern and
morphological disparity. Biology Letters, 4, 733-736.

Brusatte, S. L., Lloyd, G. T., Wang, S. C. and Norell, M. A., 2014. Gradual
assembly of avian body plan culminated in rapid rates of evolution across
dinosaur-bird transition. Current Biology, 24, 2386-2392.

Close, R. A., Friedman, M., Lloyd, G. T. and Benson, R. B. J., 2015.
Evidence for a mid-Jurassic adaptive radiation in mammals. Current Biology,
25, 2137-2142.

Lloyd, G. T., Wang, S. C. and Brusatte, S. L., 2012. Identifying
heterogeneity in rates of morphological evolution: discrete character change
in the evolution of lungfish (Sarcopterygii; Dipnoi). Evolution, 66,
330-348.
}
\keyword{disparity,distance,morphology,phylogeny}

