% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompactifyMatrix.R
\name{CompactifyMatrix}
\alias{CompactifyMatrix}
\title{Collapses matrix to unique character state distributions}
\usage{
CompactifyMatrix(clad.matrix)
}
\arguments{
\item{clad.matrix}{The cladistic matrix in the format imported by \link{ReadMorphNexus}.}
}
\description{
Collapses matrix to unique character state distributions.
}
\details{
Important: not recommended for general use.

This function is intended to make a matrix with redundant character state distributions smaller by collapsing these to single characters and upweighting them accordingly. It is intended purely for us with MRP matrices, but may have some very restricted uses elsewhere.

The function also deletes any characters weighted zero from the matrix.
}
\examples{

# Examine the matrix pre-compactification:
Michaux1989$matrix

# Examine the wieights pre-compactification:
Michaux1989$weights

# Compactify the matrix:
Michaux1989compact <- CompactifyMatrix(Michaux1989)

# Examine the matrix post-compactification:
Michaux1989compact$matrix

# Examine the wieights post-compactification:
Michaux1989compact$weights

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\seealso{
\link{SafeTaxonomicReduction} and \link{MatrixPruner}
}
\keyword{NEXUS}

