% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Claddis-package.R
\docType{data}
\name{Michaux1989}
\alias{Michaux1989}
\title{Character-taxon matrix from Michaux 1989}
\format{A character-taxon matrix in the format imported by
\link{ReadMorphNexus}.}
\description{
The character-taxon matrix from Michaux (1989).
}
\references{
Michaux, B., 1989. Cladograms can reconstruct phylogenies: an
example from the fossil record. Alcheringa, 13, 21-36.
}
\keyword{datasets}
