% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorphDistMatrix.R
\name{MorphDistMatrix}
\alias{MorphDistMatrix}
\title{Get distance matrices from a cladistic matrix}
\usage{
MorphDistMatrix(morph.matrix, distance = "MORD",
  transform.proportional.distances = "arcsine_sqrt",
  polymorphism.behaviour = "min.difference",
  uncertainty.behaviour = "min.difference")
}
\arguments{
\item{morph.matrix}{A character-taxon matrix in the format imported by \link{ReadMorphNexus}.}

\item{distance}{The distance metric to use. Must be one of \code{"GC"}, \code{"GED"}, \code{"RED"}, or \code{"MORD"} (the default).}

\item{transform.proportional.distances}{Whether to transform the proportional distances (Gower and max). Options are \code{"none"}, \code{"sqrt"}, or \code{"arcsine_sqrt"} (the default).}

\item{polymorphism.behaviour}{The distance behaviour for dealing with polymorphisms. Must be one of \code{"mean.difference"} or \code{"min.difference"} (the default.}

\item{uncertainty.behaviour}{The distance behaviour for dealing with uncertainties. Must be one of \code{"mean.difference"} or \code{"min.difference"} (the default.}
}
\value{
\item{DistanceMetric}{The distance metric used.}
\item{DistanceMatrix}{The distance matrix returned.}
\item{ComparableCharacterMatrix}{The matrix of comparable characters used for each distance.}
}
\description{
Takes a cladistic morphological dataset and converts it into a set of pairwise distances.
}
\examples{

# Get morphological distances for Day et
# al. (2016) data set:
distances <- MorphDistMatrixFast(Day2016)

# Show distance metric:
distances$DistanceMetric

# Show distance matrix:
distances$DistanceMatrix

# Show number of characters that can be scored for
# each pairwise comparison:
distances$ComparableCharacterMatrix

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com} and Thomas Guillerme \email{guillert@tcd.ie}
}
\keyword{distance}
