% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MorphMatrix2PCoA.R
\name{MorphMatrix2PCoA}
\alias{MorphMatrix2PCoA}
\title{Principal Coordinates on a Cladistic Matrix}
\usage{
MorphMatrix2PCoA(morph.matrix, distance.method = "MORD",
  transform.proportional.distances = "arcsine_sqrt",
  correction = "cailliez", tree = NULL, estimate.allchars = FALSE,
  estimate.tips = FALSE)
}
\arguments{
\item{morph.matrix}{A vector of mode character representing the tip names for which an ancestor is sought.}

\item{distance.method}{The distance method to use (one of "RED", "GED", "GC", or "MORD" - the default). See \link{MorphDistMatrix} for more details.}

\item{transform.proportional.distances}{The transformation to apply to propotional (0 to 1) distances (one of "none", "sqrt", or "arcsine_sqrt" - the default). See \link{MorphDistMatrix} for more details.}

\item{correction}{The negative eigenvalue correction to use (one of "lingoes", "none", or "cailliez" - the default). See \link{pcoa} for more details.}

\item{tree}{If a phylmorphospace is desired then a tree with root age and branch-lengths must be included.}

\item{estimate.allchars}{If including a tree whether you want to estinate ancestral states for all characters (default is FALSE). See \link {AncStateEstMatrix} for more details.}

\item{estimate.tips}{If including a tree whether you want to estinate missing or polymorphic tip states (default is FALSE). See \link {AncStateEstMatrix} for more details.}
}
\value{
\item{tree}{The tree (if supplied). Note this may be pruned from the input tree by \link{TrimMorphDistMatrix}.}

\item{dist.matrix}{The distance matrix. Note this may be pruned by \link{TrimMorphDistMatrix} and thus not include all taxa.}

\item{removed.taxa}{A vector of taxa (or nodes) removed by \link{TrimMorphDistMatrix}. Returns NULL if none are removed.}

\item{note}{See \link{pcoa}.}

\item{values}{See \link{pcoa}.}

\item{vectors}{See \link{pcoa}.}

\item{trace}{See \link{pcoa}.}

\item{vectors.cor}{See \link{pcoa}.}

\item{trace.cor}{See \link{pcoa}.}
}
\description{
Performs Principal Coordinates Analysis (PCoA) on a cladistic matrix.
}
\details{
Gower (1966). Caillez (1983). Uses \link{pcoa} from the \link{ape} package.
}
\examples{
# Nothing yet
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{coordinates}
\keyword{principal}

