% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MorphospacePlot.R
\name{MorphospacePlot}
\alias{MorphospacePlot}
\title{Plot Morphopace}
\usage{
MorphospacePlot(pcoa_input, x_axis = 1, y_axis = 2, z_axis = NULL,
  plot_taxon_names = FALSE)
}
\arguments{
\item{pcoa_input}{Text.}

\item{x_axis}{Text.}

\item{y_axis}{Text.}

\item{z_axis}{Text.}

\item{plot_taxon_names}{Text.}
}
\description{
Plots a morphospace using the output from MorphMatrix2PCoA.
}
\details{
Input must come from \link{MorphMatrix2PCoA}.
}
\examples{
# Nothing yet
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\references{
Wills? Brusatte?
}
\keyword{coordinates}
\keyword{principal}

