% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_changes_on_tree.R
\name{plot_changes_on_tree}
\alias{plot_changes_on_tree}
\title{Plots character changes on branches}
\usage{
plot_changes_on_tree(character.changes, tree)
}
\arguments{
\item{character.changes}{A matrix of character changes.}

\item{tree}{Tree on which character changes occur.}
}
\value{
A plot of character changes on a tree.
}
\description{
Plots character changes in boxes on branches.
}
\details{
Takes the \code{character.changes} output from \link{test_rates} and plots it on the tree used to generate it.
}
\examples{

# Set random seed:
set.seed(17)

# Generate a random tree for the Michaux data set:
tree <- rtree(nrow(Michaux1989$matrix_1$Matrix))

# Update taxon names to match those in the data matrix:
tree$tip.label <- rownames(Michaux1989$matrix_1$Matrix)

# Set root time by making youngest taxon extant:
tree$root.time <- max(diag(vcv(tree)))

# Get discrete character rates (includes changes):
out <- test_rates(tree, Michaux1989,
  seq(tree$root.time, 0, length.out = 3),
  BranchPartitionsToTest = list(list(1)), alpha = 0.01)

# Plot character changes on the tree:
plot_changes_on_tree(out$InferredCharacterChanges,
  tree)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
