% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cladisticMatrix.r
\name{print.cladisticMatrix}
\alias{print.cladisticMatrix}
\title{Compact display of a cladistic matrix}
\usage{
\method{print}{cladisticMatrix}(cladistic_matrix)
}
\arguments{
\item{cladistic_matrix}{An object of class \code{"cladisticMatrix"}.}
}
\value{
A summary describing the dimensions and nature of an object of class \code{"cladisticMatrix"}.
}
\description{
Displays a compact summary of the dimensions and nature of a cladistic matrix object.
}
\details{
TO ADD.
}
\examples{

# Show print.cladisticMatrix version of each included data sets:
print.cladisticMatrix(cladistic_matrix = day_2016)
print.cladisticMatrix(cladistic_matrix = gauthier_1986)
print.cladisticMatrix(cladistic_matrix = michaux_1989)

}
\seealso{
TO ADD
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
